requirements/enterprise_requirements.txt
# Enterprise-level Additional Requirements
# Add these to your existing requirements.txt

# API Documentation & Rate Limiting
flask-restx>=1.1.0
flask-limiter>=3.0.0

# Caching & Background Tasks
redis>=4.5.0
celery>=5.2.0
celery[redis]>=5.2.0

# Advanced Database & Analytics
sqlalchemy-utils>=0.38.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Monitoring & Error Tracking
sentry-sdk[flask]>=1.40.0
prometheus-client>=0.17.0

# Payment Processing
stripe>=5.5.0
paypal-checkout-serversdk>=1.0.1

# Email & Notifications
sendgrid>=6.10.0
twilio>=8.8.0

# Data Processing & Analytics
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# PDF & Report Generation
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Security & Authentication
pyjwt>=2.8.0
cryptography>=41.0.0
bcrypt>=4.0.0

# Image Processing & Storage
pillow>=10.0.0
boto3>=1.26.0  # AWS S3

# Testing & Quality
pytest-cov>=4.1.0
pytest-mock>=3.10.0
black>=23.3.0
flake8>=6.0.0

# Development Tools
python-dotenv>=1.0.0
gunicorn>=21.2.0
supervisor>=4.2.0

# API Client Libraries
requests>=2.31.0
httpx>=0.24.0

# Date & Time Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration Management
pydantic>=2.1.0
pydantic-settings>=2.0.0

# Logging & Monitoring
structlog>=23.1.0
rich>=13.4.0

# Webhook Processing
hmac>=1.0.1
hashlib>=1.0.1

# Data Validation
marshmallow>=3.19.0
email-validator>=2.0.0

# Background Job Scheduling
apscheduler>=3.10.0
celery-beat>=2.5.0

# Performance Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0
